Test: add tests for core seeder functionality and demo data preparation
Дополнительно сделан рефакторинг расположения и найминг
для интеграционных тестов Seeders
