+++ tests/Infrastructure.UnitTests/Seeding/IntegrationTestsSeeders/Shared/TestEntity.cs
@@ -1,4 +1,4 @@
-﻿namespace Infrastructure.UnitTests.Seeding.UnitTests.Shared;
+﻿namespace Infrastructure.UnitTests.Seeding.IntegrationTestsSeeders.Shared;
 
 public class TestEntity
 {
