+++ tests/Infrastructure.UnitTests/Seeding/IntegrationTestsSeeders/Shared/TestSeedCoreOptions.cs
