+++ tests/Infrastructure.UnitTests/Seeding/IntegrationTestsSeeders/Shared/TestSeederDto.cs
@@ -1,0 +1,3 @@
+﻿namespace Infrastructure.UnitTests.Seeding.IntegrationTestsSeeders.Shared;
+
+public record TestSeederDto(Guid Id, string Value);
