--- tests/Infrastructure.UnitTests/Seeding/UnitTests/Shared/TestSeederDto.cs
@@ -1,3 +1,0 @@
-﻿namespace Infrastructure.UnitTests.Seeding.UnitTests.Shared;
-
-public record TestSeederDto(Guid Id, string Value);
